Extensible Site Editor: Lift template activation restriction
