+++ lib/experiments-page.php
@@ -254,9 +254,8 @@ function gutenberg_initialize_experiments_settings() {
 		'gutenberg-experiments',
 		'gutenberg_experiments_section',
 		array(
-			'label'    => __( 'Redirects the default site editor (Appearance > Design) to use the extensible site editor page. Requires "Template Activation" experiment to be enabled.', 'gutenberg' ),
+			'label'    => __( 'Redirects the default site editor (Appearance > Design) to use the extensible site editor page.', 'gutenberg' ),
 			'id'       => 'gutenberg-extensible-site-editor',
-			'requires' => 'active_templates',
 		)
 	);
