+++ routes/template/route.ts
@@ -8,11 +8,14 @@ import { redirect } from '@wordpress/route';
  */
 export const route = {
 	beforeLoad: () => {
+		const isTemplateActivateEnabled =
+			typeof window !== 'undefined' &&
+			window.__experimentalTemplateActivate;
 		throw redirect( {
 			throw: true,
 			to: '/templates/list/$activeView',
 			params: {
-				activeView: 'active',
+				activeView: isTemplateActivateEnabled ? 'active' : 'all',
 			},
 		} );
 	},
